fix: moving components files to outside /app to fix testing on CI
