+++ frontend/src/app/401-page/page.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
 import { IoMdCloseCircleOutline } from 'react-icons/io';
-import UnLoggedHeader from '../(components)/headers/unlogged-header';
+import UnLoggedHeader from '@/components/Headers/unlogged-header';
 
 export default function NotPermittedPageWarning() {
   return (
