+++ frontend/src/app/portal/403-page/page.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@/app/(components)/ui/button';
+import { Button } from '@/components/ui/button';
 import Link from 'next/link';
 import { BsSignStop } from 'react-icons/bs';
