+++ frontend/src/app/portal/giftList/[id]/(components)/new-gift-form/new-gift-form.tsx
@@ -3,7 +3,7 @@ import { newGiftsSchema } from '@/zodSchemas/giftsSchema';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { SubmitHandler, useFieldArray, useForm } from 'react-hook-form';
 import { z } from 'zod';
-import { Spinner } from '@/app/(components)/Common/spinner/spinner';
+import { Spinner } from '@/components/Common/spinner/spinner';
 import {
   Card,
   CardContent,
@@ -11,11 +11,11 @@ import {
   CardFooter,
   CardHeader,
   CardTitle,
-} from '@/app/(components)/ui/card';
-import { Button } from '@/app/(components)/ui/button';
-import { Label } from '@/app/(components)/ui/label';
+} from '@/components/ui/card';
+import { Button } from '@/components/ui/button';
+import { Label } from '@/components/ui/label';
 import { CircleX } from 'lucide-react';
-import { Input } from '@/app/(components)/ui/input';
+import { Input } from '@/components/ui/input';
 import { useGeneralStore } from '@/stores/general/general.provider';
 import { useShallow } from 'zustand/shallow';
