+++ frontend/src/app/portal/giftsTable/[id]/(components)/data-table.tsx
@@ -17,9 +17,9 @@ import {
   TableHead,
   TableHeader,
   TableRow,
-} from '@/app/(components)/ui/table';
+} from '@/components/ui/table';
 
-import { Input } from '@/app/(components)/ui/input';
+import { Input } from '@/components/ui/input';
 import React from 'react';
 
 interface DataTableProps<TData, TValue> {
