+++ frontend/src/app/registerPage/(components)/RegistrationForms/sign-up-form.tsx
@@ -6,11 +6,11 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import Link from 'next/link';
 import { SubmitHandler, useForm } from 'react-hook-form';
 import { z } from 'zod';
-import { Spinner } from '@/app/(components)/Common/spinner/spinner';
-import { Card, CardContent, CardFooter, CardHeader, CardTitle } from '@/app/(components)/ui/card';
-import { Input } from '@/app/(components)/ui/input';
-import { Label } from '@/app/(components)/ui/label';
-import { Button } from '@/app/(components)/ui/button';
+import { Spinner } from '@/components/Common/spinner/spinner';
+import { Card, CardContent, CardFooter, CardHeader, CardTitle } from '@/components/ui/card';
+import { Input } from '@/components/ui/input';
+import { Label } from '@/components/ui/label';
+import { Button } from '@/components/ui/button';
 import { useGeneralStore } from '@/stores/general/general.provider';
 import { useShallow } from 'zustand/shallow';
