+++ frontend/src/components/Headers/logged-header.tsx
@@ -9,7 +9,7 @@ import {
   DropdownMenuLabel,
   DropdownMenuSeparator,
   DropdownMenuTrigger,
-} from '@/app/(components)/ui/dropdown-menu';
+} from '@/components/ui/dropdown-menu';
 import {
   Menu,
   LogOut,
