+++ frontend/src/components/modals/delete-modal.tsx
